Updated backend tests
